fix: maybe no more weird freezes?
this is a weird one: sometimes, after sending some changes to the state
worker, the GUI will completely freeze. No error or panic reported, just
completely frozen, must be terminated and restarted. This is super
annoying! I tried to debug it, and it seems that the main GUI thread
gets blocked inside an unsafe block of crate `parking_lot` invoking a
FUTEX syscall (Fast Userspace muTEX). Climbing up the stack trace, it
seems to be originating from accessing a watch channel, specifically
when rendering panels. I noticed that there were unnecessary borrow
calls, and slimmed them down, and still haven't experienced it again.
Which is weird! Seems like a very "magic" fix, but it is to be expected
with race conditions, and this looks to be the case. I could quite
reliably reproduce it with ~20 metrics on ~4 sources set up.

idk, I just want this fixed, but I'm still super bummed I didn't catch
the culprit...